backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
pymongo==4.7.2
motor==3.4.0
python-dotenv==1.0.1
reportlab==4.2.0
scipy==1.13.0
groq==0.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
matplotlib>=3.7.0
